test: Move existing tests of sort by tag to TagsField.test.ts
